+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Wed, 29 Jan 2025 18:28:29 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 30 Jan 2025 19:42:04 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 32.11.5
+
+Thu, 30 Jan 2025 19:42:04 GMT
+
+### Patches
+
+- Account for horizontal scrollbar when determining value of tableFitRowsHeight token ([ni/nimble@1ae9e67](https://github.com/ni/nimble/commit/1ae9e673328c3f03722ee2efd32e6d0f4ea8c6a5))
+
+## 32.11.4
+
+Thu, 30 Jan 2025 18:42:36 GMT
+
+### Patches
+
+- Fix for Select page object clickOption method. ([ni/nimble@74c078a](https://github.com/ni/nimble/commit/74c078a091d6ef4b2f80a4dcd03ff0cd168cd241))
 
 ## 32.11.3